requirements.txt
django
djangorestframework
django-unixdatetimefield
django-unfold
python-dotenv
drf-yasg
whitenoise